Release 1.4.1
V1.4.0:
Fixed bug where {MONTH_SHORT} and {MONTH_LONG} always showed the current month rather than the month corresponding to the UC.
Added buttons to view the UC of other days. To use on touch screen devices, tap the screen and either swipe or use the on-screen buttons. To use on non touch screen devices, press the enter button and use the arrow keys on the device.
V1.4.1
Fixed bug where you couldn't browse on non touch screen devices due to a lack of catching the button presses, this has been fixed by requiring the user to first press the enter key before using the arrow keys to browse.